separated the payment option from Pricing form to payment page
